move removesuffix
